+++ internal/js_lexer/js_lexer.go
@@ -18,7 +18,6 @@ import (
 	"strconv"
 	"strings"
 	"unicode"
-	"unicode/utf16"
 	"unicode/utf8"
 
 	"github.com/evanw/esbuild/internal/js_ast"
@@ -250,7 +249,7 @@ type Lexer struct {
 	// Escape sequences in string literals are decoded lazily because they are
 	// not interpreted inside tagged templates, and tagged templates can contain
 	// invalid escape sequences. If the decoded array is nil, the encoded value
-	// should be passed to "decodeEscapeSequences" first.
+	// should be passed to "tryToDecodeEscapeSequences" first.
 	decodedStringLiteralOrNil []uint16
 	encodedStringLiteralStart int
 	encodedStringLiteralText  string
@@ -320,62 +319,67 @@ func (lexer *Lexer) Raw() string {
 func (lexer *Lexer) StringLiteral() []uint16 {
 	if lexer.decodedStringLiteralOrNil == nil {
 		// Lazily decode escape sequences if needed
-		lexer.decodedStringLiteralOrNil = lexer.decodeEscapeSequences(
-			lexer.encodedStringLiteralStart,
-			lexer.encodedStringLiteralText,
-		)
+		if decoded, ok, end := lexer.tryToDecodeEscapeSequences(lexer.encodedStringLiteralStart, lexer.encodedStringLiteralText); !ok {
+			lexer.end = end
+			lexer.SyntaxError()
+		} else {
+			lexer.decodedStringLiteralOrNil = decoded
+		}
 	}
 	return lexer.decodedStringLiteralOrNil
 }
 
-func (lexer *Lexer) RawTemplateContents() string {
-	var text string
+func (lexer *Lexer) CookedAndRawTemplateContents() ([]uint16, string) {
+	var raw string
+
 	switch lexer.Token {
 	case TNoSubstitutionTemplateLiteral, TTemplateTail:
 		// "`x`" or "}x`"
-		text = lexer.source.Contents[lexer.start+1 : lexer.end-1]
+		raw = lexer.source.Contents[lexer.start+1 : lexer.end-1]
 
 	case TTemplateHead, TTemplateMiddle:
 		// "`x${" or "}x${"
-		text = lexer.source.Contents[lexer.start+1 : lexer.end-2]
+		raw = lexer.source.Contents[lexer.start+1 : lexer.end-2]
 	}
 
-	if strings.IndexByte(text, '\r') == -1 {
-		return text
-	}
+	if strings.IndexByte(raw, '\r') != -1 {
+		// From the specification:
+		//
+		// 11.8.6.1 Static Semantics: TV and TRV
+		//
+		// TV excludes the code units of LineContinuation while TRV includes
+		// them. <CR><LF> and <CR> LineTerminatorSequences are normalized to
+		// <LF> for both TV and TRV. An explicit EscapeSequence is needed to
+		// include a <CR> or <CR><LF> sequence.
 
-	// From the specification:
-	//
-	// 11.8.6.1 Static Semantics: TV and TRV
-	//
-	// TV excludes the code units of LineContinuation while TRV includes
-	// them. <CR><LF> and <CR> LineTerminatorSequences are normalized to
-	// <LF> for both TV and TRV. An explicit EscapeSequence is needed to
-	// include a <CR> or <CR><LF> sequence.
+		bytes := []byte(raw)
+		end := 0
+		i := 0
 
-	bytes := []byte(text)
-	end := 0
-	i := 0
+		for i < len(bytes) {
+			c := bytes[i]
+			i++
 
-	for i < len(bytes) {
-		c := bytes[i]
-		i++
+			if c == '\r' {
+				// Convert '\r\n' into '\n'
+				if i < len(bytes) && bytes[i] == '\n' {
+					i++
+				}
 
-		if c == '\r' {
-			// Convert '\r\n' into '\n'
-			if i < len(bytes) && bytes[i] == '\n' {
-				i++
+				// Convert '\r' into '\n'
+				c = '\n'
 			}
 
-			// Convert '\r' into '\n'
-			c = '\n'
+			bytes[end] = c
+			end++
 		}
 
-		bytes[end] = c
-		end++
+		raw = string(bytes[:end])
 	}
 
-	return string(bytes[:end])
+	// This will return nil on failure, which will become "undefined" for the tag
+	cooked, _, _ := lexer.tryToDecodeEscapeSequences(lexer.start+1, raw)
+	return cooked, raw
 }
 
 func (lexer *Lexer) IsIdentifierOrKeyword() bool {
@@ -1695,7 +1699,12 @@ func (lexer *Lexer) scanIdentifierWithEscapes(kind identifierKind) (string, T) {
 	}
 
 	// Second pass: re-use our existing escape sequence parser
-	text := string(utf16.Decode(lexer.decodeEscapeSequences(lexer.start, lexer.Raw())))
+	decoded, ok, end := lexer.tryToDecodeEscapeSequences(lexer.start, lexer.Raw())
+	if !ok {
+		lexer.end = end
+		lexer.SyntaxError()
+	}
+	text := string(UTF16ToString(decoded))
 
 	// Even though it was escaped, it must still be a valid identifier
 	identifier := text
@@ -2183,7 +2192,9 @@ func fixWhitespaceAndDecodeJSXEntities(text string) []uint16 {
 	return decoded
 }
 
-func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
+// If this fails, this returns "nil, false, end" where "end" is the value to
+// store to "lexer.end" before calling "lexer.SyntaxError()" if relevant
+func (lexer *Lexer) tryToDecodeEscapeSequences(start int, text string) ([]uint16, bool, int) {
 	decoded := []uint16{}
 	i := 0
 
@@ -2238,8 +2249,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 
 			case 'v':
 				if lexer.json.parse {
-					lexer.end = start + i - width2
-					lexer.SyntaxError()
+					return nil, false, start + i - width2
 				}
 
 				decoded = append(decoded, '\v')
@@ -2248,8 +2258,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 			case '0', '1', '2', '3', '4', '5', '6', '7':
 				octalStart := i - 2
 				if lexer.json.parse {
-					lexer.end = start + i - width2
-					lexer.SyntaxError()
+					return nil, false, start + i - width2
 				}
 
 				// 1-3 digit octal
@@ -2289,8 +2298,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 
 			case 'x':
 				if lexer.json.parse {
-					lexer.end = start + i - width2
-					lexer.SyntaxError()
+					return nil, false, start + i - width2
 				}
 
 				// 2-digit hexadecimal
@@ -2306,8 +2314,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 					case 'A', 'B', 'C', 'D', 'E', 'F':
 						value = value*16 | (c3 + 10 - 'A')
 					default:
-						lexer.end = start + i - width3
-						lexer.SyntaxError()
+						return nil, false, start + i - width3
 					}
 				}
 				c = value
@@ -2322,8 +2329,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 
 				if c3 == '{' {
 					if lexer.json.parse {
-						lexer.end = start + i - width2
-						lexer.SyntaxError()
+						return nil, false, start + i - width2
 					}
 
 					// Variable-length
@@ -2344,13 +2350,11 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 							value = value*16 | (c3 + 10 - 'A')
 						case '}':
 							if isFirst {
-								lexer.end = start + i - width3
-								lexer.SyntaxError()
+								return nil, false, start + i - width3
 							}
 							break variableLength
 						default:
-							lexer.end = start + i - width3
-							lexer.SyntaxError()
+							return nil, false, start + i - width3
 						}
 
 						if value > utf8.MaxRune {
@@ -2376,8 +2380,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 						case 'A', 'B', 'C', 'D', 'E', 'F':
 							value = value*16 | (c3 + 10 - 'A')
 						default:
-							lexer.end = start + i - width3
-							lexer.SyntaxError()
+							return nil, false, start + i - width3
 						}
 
 						if j < 3 {
@@ -2390,8 +2393,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 
 			case '\r':
 				if lexer.json.parse {
-					lexer.end = start + i - width2
-					lexer.SyntaxError()
+					return nil, false, start + i - width2
 				}
 
 				// Ignore line continuations. A line continuation is not an escaped newline.
@@ -2403,8 +2405,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 
 			case '\n', '\u2028', '\u2029':
 				if lexer.json.parse {
-					lexer.end = start + i - width2
-					lexer.SyntaxError()
+					return nil, false, start + i - width2
 				}
 
 				// Ignore line continuations. A line continuation is not an escaped newline.
@@ -2416,8 +2417,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 					case '"', '\\', '/':
 
 					default:
-						lexer.end = start + i - width2
-						lexer.SyntaxError()
+						return nil, false, start + i - width2
 					}
 				}
 
@@ -2433,7 +2433,7 @@ func (lexer *Lexer) decodeEscapeSequences(start int, text string) []uint16 {
 		}
 	}
 
-	return decoded
+	return decoded, true, 0
 }
 
 func (lexer *Lexer) RescanCloseBraceAsTemplateToken() {